+++ src/main/java/com/potaton/potato/domain/daily/repository/DailyQuestionJpaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DailyQuestionJpaRepository extends JpaRepository<DailyQuestion, Integer> {
+public interface DailyQuestionJpaRepository extends JpaRepository<DailyQuestion, Long> {
 }
