+++ src/main/java/com/potaton/potato/domain/daily/repository/QuestionJpaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface QuestionJpaRepository extends JpaRepository<Question, Integer> {
+public interface QuestionJpaRepository extends JpaRepository<Question, Long> {
 }
